Added or clause to check for valid timer
The if-statement that was checking to see if playSound(modulator) should be called was only checking for the stopTimer to be nil. This is wrong, because after the first time that a sound is played, stopTimer doesnt turn nil; it just gets invalidated. I added an or-clause that checks for an invalid timer, signifying that playSound is okay to run